pyproject.toml
[project]
name = "to-do-list"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.2.1",
]

[dependency-groups]
dev = [
    "basedpyright>=1.29.2",
    "isort>=6.0.1",
    "ruff>=0.11.11",
]

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
skip = [".bzr", ".direnv", ".eggs", ".git", ".hg", ".mypy_cache", ".nox", ".pants.d", ".svn", ".tox", ".venv", "__pypackages__", "_build", "buck-out", "build", "dist", "node_modules", "venv", "dispie/"]

[tool.ruff]
line-length = 79
lint.extend-select = ["ANN"]
lint.ignore = ["ANN401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
docstring-code-line-length = 72

[tool.basedpyright]
exclude = ["**/node_modules", "**/__pycache__", "**/.*", "test*.py"]
reportImportCycles = false
reportUnnecessaryIsInstance = false
reportImplicitStringConcatenation = false
reportUnusedCallResult = false
reportMissingTypeStubs = false
reportAssignmentType = false
reportRedeclaration = false
reportAny = false
reportExplicitAny = false
reportUnusedParameter = false
